+++ src/components/Footer/Categories/Categories.module.scss
@@ -1,0 +1,38 @@
+@import '@/styles/fonts';
+@import '@/styles/variables';
+
+.linkBlock {
+  display: flex;
+  align-items: center;
+  width: 100%;
+  margin: 0 24px;
+  border-bottom: 2px solid #e2e4f0;
+
+  .popular {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    padding: 48px 0;
+    gap: 56px;
+
+    & .itemPopular {
+      display: flex;
+      flex-direction: column;
+      font-family: 'Gotham', sans-serif;
+      font-size: 16px;
+      font-weight: 700;
+      gap: 24px;
+      & .itemLink {
+        display: flex;
+        flex-direction: column;
+        gap: 16px;
+        & span {
+          font-family: 'Gotham Light', sans-serif;
+          font-size: 14px;
+          font-weight: 400;
+          color: #050d23;
+        }
+      }
+    }
+  }
+}
