+++ src/components/Footer/Footer.module.scss
@@ -8,162 +8,4 @@
   width: 100%;
   background: #ffffff;
   gap: 32px;
-
-  .contactsBlock {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-    margin: 0 24px;
-    border-top: 2px solid #e2e4f0;
-    border-bottom: 2px solid #e2e4f0;
-
-    .contacts {
-      display: flex;
-      padding: 48px 0;
-      gap: 56px;
-
-      & img {
-        width: 206px;
-        height: 40px;
-      }
-
-      .itemInfo {
-        display: flex;
-        flex-direction: column;
-        justify-content: space-between;
-
-        .itemInfoTitle {
-          font-family: 'Gotham', sans-serif;
-          font-size: 16px;
-          font-weight: 700;
-        }
-
-        & span {
-          font-family: 'Gotham', sans-serif;
-          font-size: 12px;
-          font-weight: 400;
-          color: $subtextColor;
-        }
-      }
-    }
-
-    .popular {
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      gap: 32px;
-
-      & span {
-        font-family: 'Gotham', sans-serif;
-        font-size: 16px;
-        font-weight: 700;
-      }
-
-      .idea {
-        display: flex;
-        width: 179px;
-        height: 48px;
-        justify-content: center;
-        align-items: center;
-        background: #f2f6ff;
-        border-radius: 5px;
-        color: $buttonColor;
-      }
-    }
-  }
-
-  .linkBlock {
-    display: flex;
-    align-items: center;
-    width: 100%;
-    margin: 0 24px;
-    border-bottom: 2px solid #e2e4f0;
-
-    .popular {
-      display: flex;
-      justify-content: space-between;
-      width: 100%;
-      padding: 48px 0;
-      gap: 56px;
-
-      & .itemPopular {
-        display: flex;
-        flex-direction: column;
-        font-family: 'Gotham', sans-serif;
-        font-size: 16px;
-        font-weight: 700;
-        gap: 24px;
-        & .itemLink {
-          display: flex;
-          flex-direction: column;
-          gap: 16px;
-          & span {
-            font-family: 'Gotham Light', sans-serif;
-            font-size: 14px;
-            font-weight: 300;
-          }
-        }
-      }
-    }
-  }
-  .basement {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-between;
-    margin: 40px 24px;
-    gap: 40px;
-
-    .social {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      gap: 28px;
-      .socialItem {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        gap: 8px;
-
-        & img {
-          width: 16px;
-          height: 16px;
-        }
-
-        & span {
-          font-family: 'Gotham', sans-serif;
-          font-size: 12px;
-          font-weight: 400;
-          color: $subtextColor;
-        }
-      }
-      & :nth-child(4) span {
-        color: $buttonColor;
-      }
-    }
-    .politic {
-      text-align: center;
-      font-family: 'Gotham Light', sans-serif;
-      font-size: 12px;
-      font-weight: 300;
-      color: $subtextColor;
-    }
-    .logoFooter {
-      display: flex;
-      align-items: center;
-      gap: 24px;
-      & img {
-        width: 124px;
-        height: 24px;
-      }
-
-      & span {
-        font-family: 'Gotham', sans-serif;
-        font-size: 14px;
-        font-weight: 700;
-        color: $subtextColor;
-      }
-    }
-  }
 }
