+++ src/components/Header/Header.module.scss
@@ -1,5 +1,5 @@
-@import '../../styles/fonts';
-@import '../../styles/variables';
+@import '@/styles/fonts';
+@import '@/styles/variables';
 
 .header {
   display: flex;
@@ -8,168 +8,11 @@
   height: 136px;
   justify-content: space-between;
   margin-bottom: 10px;
-
-  .logo {
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    width: 100%;
-    height: 80px;
-    padding: 16px 24px 16px 24px;
-    background: #ffffff;
-    margin-bottom: 5px;
-    gap: 24px;
-    & img {
-      width: 144px;
-      height: 28px;
-    }
-    & button {
-      width: 128px;
-      height: 48px;
-      font-family: 'Gotham', sans-serif;
-      text-transform: capitalize;
-      font-size: 14px;
-      font-weight: bold;
-      gap: 12px;
-      padding: 16px 24px 16px 24px;
-      background: $buttonColor;
-
-      & svg {
-        font-size: 24px;
-      }
-    }
-    .search {
-      display: flex;
-      align-items: center;
-      width: 730px;
-
-      input,
-      label {
-        width: 100%;
-        height: 48px;
-        color: $textColor;
-        font-family: 'Gotham Pro', sans-serif;
-        font-weight: 400;
-        font-size: 14px;
-        box-sizing: border-box;
-        margin: 0;
-      }
-    }
-    .iconsHeader {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      gap: 24px;
-      & hr {
-        height: 48px;
-        background: #e2e4f0;
-      }
-
-      & img {
-        width: 32px;
-        height: 32px;
-      }
-    }
-    .basket {
-      display: flex;
-      justify-content: space-evenly;
-      align-items: center;
-      width: 148px;
-      height: 48px;
-      background: #f2f6ff;
-      border-radius: 5px;
-
-      & img {
-        width: 24px;
-        height: 24px;
-      }
-      & span {
-        color: $buttonColor;
-        font-family: 'Gotham', sans-serif;
-        font-size: 14px;
-        font-weight: bold;
-      }
-    }
-    .profile {
-      display: flex;
-      width: 206px;
-      justify-content: space-evenly;
-      align-items: center;
-      gap: 16px;
-      & img {
-        width: 32px;
-        height: 32px;
-      }
-      & span {
-        font-family: 'Gotham Pro', sans-serif;
-        font-size: 14px;
-        font-weight: 400;
-      }
-      & select {
-        width: 3px;
-        border: none;
-        font-family: 'Gotham Pro', sans-serif;
-        font-size: 14px;
-        background: right 0.8em top 50% no-repeat
-          url('../../../src/styles/images/svg/arrow_down.svg');
-        padding-right: 36px;
-        appearance: none;
-      }
-    }
-  }
-  .navigation {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    width: 100%;
-    height: 56px;
-    padding: 16px 48px 16px 26px;
-    background: #ffffff;
-    //margin-bottom: 5px;
-    gap: 24px;
-
-    .nav {
-      display: flex;
-      align-items: center;
-      gap: 24px;
-
-      & span {
-        font-family: 'Gotham Pro', sans-serif;
-        font-size: 14px;
-        font-weight: normal;
-      }
-    }
-
-    .optionBlock {
-      display: flex;
-      align-items: center;
-      width: 321px;
-      height: 32px;
-      color: $buttonColor;
-      font-family: 'Gotham Pro', sans-serif;
-      font-size: 14px;
-      gap: 25px;
-
-      .message {
-        display: flex;
-        justify-content: space-evenly;
-        align-items: center;
-        width: 159px;
-        height: 32px;
-        background: #f2f6ff;
-        border-radius: 5px;
-
-        & img {
-          width: 16px;
-          height: 16px;
-        }
-        & span {
-          color: $buttonColor;
-          font-family: 'Gotham Pro', sans-serif;
-          font-size: 14px;
-        }
-      }
-    }
+}
+@media (max-width: 1240px) {
+  .header {
+    flex-wrap: wrap;
+    height: fit-content;
+    gap: 8px;
   }
 }
