+++ src/components/Header/Logo/LogoString.module.scss
@@ -1,0 +1,76 @@
+@import '@/styles/fonts';
+@import '@/styles/variables';
+
+.logo {
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  width: 100%;
+  height: 80px;
+  padding: 16px 24px 16px 24px;
+  background: #ffffff;
+  margin-bottom: 5px;
+  gap: 24px;
+  & img {
+    width: 144px;
+    height: 28px;
+  }
+  & button {
+    width: 128px;
+    height: 48px;
+    font-family: 'Gotham', sans-serif;
+    text-transform: capitalize;
+    font-size: 14px;
+    font-weight: bold;
+    gap: 12px;
+    padding: 16px 24px 16px 24px;
+    background: $buttonColor;
+
+    & svg {
+      font-size: 24px;
+    }
+  }
+
+  .iconsHeader {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 24px;
+    & hr {
+      height: 48px;
+      background: #e2e4f0;
+    }
+
+    & img {
+      width: 32px;
+      height: 32px;
+    }
+  }
+  .basket {
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    width: 148px;
+    height: 48px;
+    background: #f2f6ff;
+    border-radius: 5px;
+
+    & img {
+      width: 24px;
+      height: 24px;
+    }
+    & span {
+      color: $buttonColor;
+      font-family: 'Gotham', sans-serif;
+      font-size: 14px;
+      font-weight: bold;
+    }
+  }
+}
+@media (max-width: 1240px) {
+  .logo {
+    flex-wrap: wrap;
+    height: fit-content;
+    gap: 8px;
+  }
+}
