+++ src/components/Header/Logo/Profile/Profile.module.scss
@@ -1,0 +1,28 @@
+@import '@/styles/fonts';
+@import '@/styles/variables';
+
+.profile {
+  display: flex;
+  width: 206px;
+  justify-content: space-evenly;
+  align-items: center;
+  & img {
+    width: 32px;
+    height: 32px;
+  }
+  & span {
+    width: 160px;
+    font-family: 'Gotham Pro', sans-serif;
+    font-size: 14px;
+    font-weight: 400;
+  }
+  & select {
+    width: 3px;
+    border: none;
+    font-family: 'Gotham Pro', sans-serif;
+    font-size: 14px;
+    background: right 0.8em top 50% no-repeat url('@/styles/images/svg/arrow_down.svg');
+    padding-right: 36px;
+    appearance: none;
+  }
+}
