+++ src/components/Header/Logo/Search/Search.module.scss
@@ -1,0 +1,20 @@
+@import '@/styles/fonts';
+@import '@/styles/variables';
+
+.search {
+  display: flex;
+  align-items: center;
+  width: 730px;
+
+  input,
+  label {
+    width: 100%;
+    height: 48px;
+    color: $textColor;
+    font-family: 'Gotham Pro', sans-serif;
+    font-weight: 400;
+    font-size: 14px;
+    box-sizing: border-box;
+    margin: 0;
+  }
+}
