+++ src/components/Header/Navigation/Navigation.module.scss
@@ -1,0 +1,64 @@
+@import '@/styles/fonts';
+@import '@/styles/variables';
+
+.navigation {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+
+  width: 100%;
+  height: 56px;
+  padding: 16px 48px 16px 26px;
+  background: #ffffff;
+  gap: 24px;
+
+  .nav {
+    display: flex;
+    align-items: center;
+    gap: 24px;
+
+    & span {
+      font-family: 'Gotham Pro', sans-serif;
+      font-size: 14px;
+      font-weight: normal;
+    }
+  }
+
+  .optionBlock {
+    display: flex;
+    align-items: center;
+    width: 321px;
+    height: 32px;
+    color: $buttonColor;
+    font-family: 'Gotham Pro', sans-serif;
+    font-size: 14px;
+    gap: 25px;
+
+    .message {
+      display: flex;
+      justify-content: space-evenly;
+      align-items: center;
+      width: 159px;
+      height: 32px;
+      background: #f2f6ff;
+      border-radius: 5px;
+
+      & img {
+        width: 16px;
+        height: 16px;
+      }
+      & span {
+        color: $buttonColor;
+        font-family: 'Gotham Pro', sans-serif;
+        font-size: 14px;
+      }
+    }
+  }
+}
+@media (max-width: 1240px) {
+  .navigation {
+    flex-wrap: wrap;
+    height: fit-content;
+    gap: 8px;
+  }
+}
