+++ src/components/PhotoBlock/PhotoBlock.module.scss
@@ -1,0 +1,22 @@
+@import '@/styles/fonts';
+@import '@/styles/variables';
+
+.photoBlock {
+  display: flex;
+  flex-direction: column;
+  width: 400px;
+  gap: 24px;
+  & img {
+    height: 400px;
+  }
+  .slider {
+    display: flex;
+    justify-content: space-between;
+
+    width: 400px;
+    height: 64px;
+    & img {
+      height: 64px;
+    }
+  }
+}
