+++ src/components/Titles/Title.tsx
@@ -1,0 +1,12 @@
+import style from './Title.module.scss'
+
+const Title = () => {
+  return (
+    <div className={style.titleBlock}>
+      <div className={style.catalogWay}>Каталог / Обувь / Кроссовки / Беговые</div>
+      <div className={style.title}>Кроссовки мужские Skechers Sunny Dale</div>
+    </div>
+  )
+}
+
+export default Title
