--- src/pages/MainPage.module.scss
@@ -1,241 +1,0 @@
-@import '@/styles/fonts';
-@import '@/styles/variables';
-
-.mainPage {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  width: 100%;
-  padding: 24px 32px;
-  background: #ffffff;
-  gap: 32px;
-
-  .titleBlock {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    width: 100%;
-    height: 62px;
-    gap: 24px;
-    .catalogWay {
-      font-family: 'Gotham', sans-serif;
-      font-size: 12px;
-      font-weight: 400;
-      color: $subtextColor;
-    }
-    .title {
-      font-family: 'Gotham Pro', sans-serif;
-      font-size: 32px;
-      font-weight: 400;
-      color: $textColor;
-      text-transform: uppercase;
-    }
-  }
-  .product {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 48px;
-
-    .photoBlock {
-      display: flex;
-      flex-direction: column;
-      width: 400px;
-      gap: 24px;
-      & img {
-        height: 400px;
-      }
-      .slider {
-        display: flex;
-        justify-content: space-between;
-
-        width: 400px;
-        height: 64px;
-        & img {
-          height: 64px;
-        }
-      }
-    }
-    .contentBlock {
-      display: flex;
-      flex-direction: column;
-      width: 70%;
-      gap: 24px;
-
-      .productDataBlock {
-        display: flex;
-        gap: 48px;
-        .prices {
-          display: flex;
-          width: 372px;
-          flex-direction: column;
-          font-family: 'Gotham', sans-serif;
-          font-size: 14px;
-          font-weight: 500;
-          gap: 32px;
-          & span {
-            color: #ff3d71;
-          }
-
-          .priceItem {
-            display: flex;
-            align-items: center;
-            gap: 16px;
-
-            & span {
-              font-family: 'Gotham', sans-serif;
-              font-size: 32px;
-              font-weight: 700;
-              color: $textColor;
-            }
-
-            .actions {
-              font-family: 'Gotham', sans-serif;
-              font-size: 16px;
-              font-weight: 700;
-              color: #ffffff;
-              background: #ff3d71;
-              padding: 5px 10px;
-              border-radius: 5px;
-            }
-          }
-          .package {
-            display: flex;
-            gap: 16px;
-            padding-bottom: 32px;
-            border-bottom: 2px solid #e2e4f0;
-            .order {
-              display: flex;
-              align-items: center;
-              gap: 12px;
-              & span {
-                color: $buttonColor;
-              }
-              .message {
-                display: flex;
-                justify-content: space-evenly;
-                align-items: center;
-                height: 26px;
-                background: #f2f6ff;
-                border-radius: 5px;
-
-                & span {
-                  color: $subtextColor;
-                  font-family: 'Gotham Pro', sans-serif;
-                  font-size: 14px;
-                  background: #f2f6ff;
-                  border-radius: 5px;
-                  padding: 5px 10px;
-                }
-              }
-            }
-          }
-          .infoBlock {
-            display: flex;
-            height: 40px;
-            justify-content: space-between;
-            .itemInfo {
-              display: flex;
-              flex-direction: column;
-
-              justify-content: space-between;
-              .itemInfoTitle {
-                font-family: 'Gotham', sans-serif;
-                font-size: 14px;
-                font-weight: 700;
-              }
-              & span {
-                font-family: 'Gotham', sans-serif;
-                font-size: 12px;
-                font-weight: 400;
-                color: $subtextColor;
-              }
-            }
-          }
-          .btn {
-            display: flex;
-            justify-content: space-between;
-            .counter {
-              display: flex;
-              justify-content: space-between;
-              align-items: center;
-              width: 284px;
-              height: 56px;
-              border: 2px solid $buttonColor;
-              border-radius: 5px;
-              padding: 16px;
-              font-family: 'Gotham', sans-serif;
-              font-size: 16px;
-              font-weight: 400;
-            }
-            & button {
-              width: 284px;
-              height: 56px;
-              font-family: 'Gotham', sans-serif;
-              font-size: 14px;
-              font-weight: 500;
-
-              & img {
-                padding-right: 16px;
-                font-weight: bold;
-              }
-            }
-            .iconHeart {
-              display: flex;
-              width: 72px;
-              height: 56px;
-              justify-content: center;
-              align-items: center;
-              background: #f2f6ff;
-              border-radius: 5px;
-            }
-          }
-        }
-        .characteristics {
-          display: flex;
-          flex-wrap: wrap;
-          width: 636px;
-          gap: 32px;
-          & h4 {
-            width: 100%;
-            font-family: 'Gotham', sans-serif;
-            font-size: 20px;
-            font-weight: 500;
-            color: #071435;
-            margin: 0;
-          }
-          & .itemInfo {
-            display: flex;
-            flex-direction: column;
-            width: 45%;
-            gap: 32px;
-            & .temInfoTitle {
-              font-family: 'Gotham', sans-serif;
-              font-size: 14px;
-              font-weight: 700;
-              color: #071435;
-            }
-            & span {
-              font-family: 'Gotham', sans-serif;
-              font-size: 12px;
-              font-weight: 400;
-              color: $subtextColor;
-            }
-          }
-        }
-      }
-      .description {
-        font-family: 'Gotham Light', sans-serif;
-        font-size: 18px;
-        font-weight: 300;
-        color: #071435;
-        padding-right: 24px;
-
-        & h3 {
-          font-family: 'Gotham', sans-serif;
-          font-weight: 500;
-          font-size: 24px;
-        }
-      }
-    }
-  }
-}
