+++ src/pages/MainPage/Description/Description.tsx
@@ -1,0 +1,20 @@
+import style from './Description.module.scss'
+
+const Description = () => {
+  return (
+    <div className={style.description}>
+      <h3>Описание товара</h3>
+      <p>
+        Создание приверженного покупателя специфицирует неопровержимый комплексный анализ ситуации.
+        CTR существенно индуцирует из ряда вон выходящий SWOT-анализ. Воздействие на потребителя
+        определяет возрастающий интеграл по поверхности, что известно даже школьникам. Отсюда
+        естественно следует, что коммуникация уравновешивает косвенный фактор коммуникации. Поле
+        направлений естественно допускает экспериментальный скачок функции, таким образом сбылась
+        мечта идиота - утверждение полностью доказано. Арифметическая прогрессия притягивает линейно
+        зависимый пул лояльных изданий.
+      </p>
+    </div>
+  )
+}
+
+export default Description
